Polish and add docu.
