--- examples/testing-with-neo4j-harness/src/test/java/org/neo4j/doc/driver/springframework/boot/simple/MovieServicePureSpringTest.java
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2019-2020 "Neo4j,"
- * Neo4j Sweden AB [https://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.doc.driver.springframework.boot.simple;
-
-import static org.assertj.core.api.Assertions.*;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.neo4j.driver.springframework.boot.test.autoconfigure.EnableNeo4jTestHarness;
-import org.neo4j.harness.Neo4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-/**
- * Example on how one could use {@link EnableNeo4jTestHarness @EnableNeo4jTestHarness}.
- *
- * @author Michael J. Simons
- */
-@SpringBootTest
-@EnableNeo4jTestHarness
-public class MovieServicePureSpringTest {
-
-	@Autowired
-	private MoviesService moviesService;
-
-	@BeforeEach
-	void prepareDatabase(@Autowired Neo4j neo4j) {
-		neo4j.defaultDatabaseService().executeTransactionally(""
-			+ "CREATE (TheMatrix:Movie {title:'The Matrix', released:1999, tagline:'Welcome to the Real World'})\n"
-			+ "CREATE (TheMatrixReloaded:Movie {title:'The Matrix Reloaded', released:2003, tagline:'Free your mind'})\n"
-			+ "CREATE (TheMatrixRevolutions:Movie {title:'The Matrix Revolutions', released:2003, tagline:'Everything that has a beginning has an end'})\n"
-		);
-	}
-
-	@Test
-	void shouldRetrieveMovieTitles() {
-
-		assertThat(moviesService.getMovieTitles())
-			.hasSize(3)
-			.contains("The Matrix");
-	}
-}
