+++ examples/testing-with-neo4j-harness/src/test/java/org/neo4j/doc/driver/springframework/boot/simple/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package includes three alternative versions that might be used in testing scenarios that need the embedded Neo4j test harness.
+ */
+package org.neo4j.doc.driver.springframework.boot.simple;
