+++ neo4j-java-driver-spring-boot-test-harness-4x-support/src/main/java/org/neo4j/driver/springframework/boot/test/autoconfigure/Neo4jTestHarnessRegistrar.java
@@ -55,7 +55,9 @@ final class Neo4jTestHarnessRegistrar implements ImportBeanDefinitionRegistrar {
 
 		@Override
 		public Neo4j getObject() {
-			return Neo4jBuilders.newInProcessBuilder().build();
+			return Neo4jBuilders.newInProcessBuilder()
+				.withDisabledServer()
+				.build();
 		}
 
 		@Override
